requirements.txt
tensorflow>=2.0.0
numpy>=1.19.2
pillow>=8.0.0
matplotlib>=3.3.2